FILE: moves combo-analyzer
