+++ libs/spirit-islander/util/src/lib/app/combo-analyzer.spec.ts
